Add files via upload
Whether in statistical learning or machine learning, it is essential to select the relevant
features. This report therefore focuses on the presentation, evaluation, and comparison of procedures
that can be used for this selection. It is based on selection methods specific to statistical
learning (procedures with inference) and others specific to machine learning (procedures without
inference). It presents these different techniques tested on simulated data, according to various
stopping criteria or conditions. In our data generation, we consider different scenarios: independence
(null hypothesis), dependence (alternative hypothesis), including situations of linear
dependence, structural break, and multi-colinearity. Outliers are also taken into account.
This paper also aims to provide a detailed analysis of feature selection methods in statistical
learning and machine learning, to better understand the strengths and weaknesses of each
approach. This will also help to guide experts in their choice of methods best fitted to their
specific needs. This study reveals the key outcomes that follow : machine learning selection
methods generally surpass statistical learning methods in terms of performance. Nevertheless,
the performance of these methods is compromised when outliers are present, and in a few rare
cases when structural breaks are introduced. This study also shows that machine learning processes
are much more affected when coefficients levels vary. This also happens when the noise
inserted in the model moves. This susceptibility to variations is most apparent when there is
correlation.

Keywords : Machine Learning, Statistical Learning, Feature selection, Lasso, Lars, Elasticnet,
Forward, Backward, Stepwise, Bootstrap, K-Fold, Cross-Validation.